requirements.txt
meshcore>=0.1.0
paho-mqtt>=1.6.1
geojson>=2.5.0
requests>=2.28.0
python-dotenv>=0.21.0
meshtastic>=2.3.0
numpy>=1.21.0
msgpack>=1.0.0
